qol: can rewrite HMs and bigger repels + harder ferry teams

+++ src/item.c
@@ -892,7 +892,7 @@ u8 GetItemHoldEffect(u16 itemId)
     return gItems[SanitizeItemId(itemId)].holdEffect;
 }
 
-u8 GetItemHoldEffectParam(u16 itemId)
+u16 GetItemHoldEffectParam(u16 itemId)
 {
     return gItems[SanitizeItemId(itemId)].holdEffectParam;
 }
